Reorganised into separate sourcefiles.
